fully implemented Extractor, some code cleanup
Working modules: ResultEditor, Decompiler and Extractor
I've tried implementing the Extractor using BCEL to be able
to use a LabTests.class file directly, but the creators stored some
important info in non final vars, couldn't use java reflection cause
the goal is to extract info from an isolated LabTetsts.class file.
AFAIK this is the only way to do this, so sticking to this dirty thing
of reading raw java code.